+++ SampleScripts/iam/iamSolutionInteroperability.ps1
@@ -10,7 +10,12 @@
     Date:   2022-10-10
     Copyright 2021-2022 VMware, Inc.
     ===================================================================================================================
-    
+    .CHANGE_LOG
+
+    - 1.1.000   (Gary Blake / 2023-07-25)   - Added Support for VCF 5.0.x Planning and Prep Workbook
+
+    ===================================================================================================================
+
     .SYNOPSIS
     Configure Solution Interoperability for Identity and Access Management
 
@@ -53,7 +58,7 @@ Try {
             Write-LogMessage -type INFO -message "Opening the Excel Workbook: $Workbook"
             $pnpWorkbook = Open-ExcelPackage -Path $Workbook
             Write-LogMessage -type INFO -message "Checking Valid Planning and Prepatation Workbook Provided"
-            if (($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.3.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.4.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.5.x")) {
+            if (($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.3.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.4.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.5.x" ) -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v5.0.x" )) {
                 Write-LogMessage -type INFO -message "Planning and Prepatation Workbook Provided Not Supported" -colour Red 
                 Break
             }